LocalYoutubeAIHelper-Python/requirements.txt
openai==1.50.0
requests==2.32.3
yt-dlp==2024.8.6
srt==3.5.1
genson==1.3.0
python-dotenv==1.0.1

# Google API related libraries
google-auth==2.23.4
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.1.0
google-api-python-client==2.101.0